Implemented Locale and used Locale#t in the existing code.
I also implemented the mock for testing against Locale. It will be useful in the future.

For now the Locale is passed to the views manually. We were considering adding some
views factory, but calling editor.view( ViewClass, model, ...other args ) would not be that
simpler and there's a problem with arguments order, because a view may accept more than
a model and then the locale must be injected in the middle of (model, ..., other args).